refactor: seperate event logger and topology reconciler into own files
