+++ controllers/state_of_the_world.go
@@ -1,24 +1,18 @@
 package controllers
 
 import (
-	"context"
 	"fmt"
-	"strings"
-	"sync"
 
 	certmanagerv1 "github.com/cert-manager/cert-manager/pkg/apis/certmanager/v1"
 	egv1alpha1 "github.com/envoyproxy/gateway/api/v1alpha1"
 	"github.com/go-logr/logr"
-	"github.com/google/go-cmp/cmp"
 	authorinov1beta1 "github.com/kuadrant/authorino-operator/api/v1beta1"
 	limitadorv1alpha1 "github.com/kuadrant/limitador-operator/api/v1alpha1"
 	"github.com/kuadrant/policy-machinery/controller"
-	"github.com/kuadrant/policy-machinery/machinery"
 	istioclientgoextensionv1alpha1 "istio.io/client-go/pkg/apis/extensions/v1alpha1"
 	istioclientnetworkingv1alpha3 "istio.io/client-go/pkg/apis/networking/v1alpha3"
 	istioclientgosecurityv1beta1 "istio.io/client-go/pkg/apis/security/v1beta1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/dynamic"
@@ -156,103 +150,6 @@ func buildReconciler(client *dynamic.DynamicClient) controller.ReconcileFunc {
 		},
 	}
 	return reconciler.Run
-}
-
-type TopologyFileReconciler struct {
-	Client    *dynamic.DynamicClient
-	Namespace string
-}
-
-func NewTopologyFileReconciler(client *dynamic.DynamicClient, namespace string) *TopologyFileReconciler {
-	if namespace == "" {
-		panic("namespace must be specified and can not be a blank string")
-	}
-	return &TopologyFileReconciler{Client: client, Namespace: namespace}
-}
-
-func (r *TopologyFileReconciler) Reconcile(ctx context.Context, _ []controller.ResourceEvent, topology *machinery.Topology, _ error, _ *sync.Map) error {
-	logger := controller.LoggerFromContext(ctx).WithName("topology file")
-
-	cm := &corev1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "topology",
-			Namespace: r.Namespace,
-			Labels:    map[string]string{kuadrant.TopologyLabel: "true"},
-		},
-		Data: map[string]string{
-			"topology": topology.ToDot(),
-		},
-	}
-	unstructuredCM, err := controller.Destruct(cm)
-	if err != nil {
-		logger.Error(err, "failed to destruct topology configmap")
-		return err
-	}
-
-	existingTopologyConfigMaps := topology.Objects().Items(func(object machinery.Object) bool {
-		return object.GetName() == cm.GetName() && object.GetNamespace() == cm.GetNamespace() && object.GroupVersionKind().Kind == ConfigMapGroupKind.Kind
-	})
-
-	if len(existingTopologyConfigMaps) == 0 {
-		_, err = r.Client.Resource(controller.ConfigMapsResource).Namespace(cm.Namespace).Create(ctx, unstructuredCM, metav1.CreateOptions{})
-		if err != nil {
-			if errors.IsAlreadyExists(err) {
-				// This error can happen when the operator is starting, and the create event for the topology has not being processed.
-				logger.Info("already created topology configmap, must not be in topology yet")
-				return err
-			}
-			logger.Error(err, "failed to write topology configmap")
-		}
-		return err
-	}
-
-	if len(existingTopologyConfigMaps) > 1 {
-		logger.Info("multiple topology configmaps found, continuing but unexpected behaviour may occur")
-	}
-	existingTopologyConfigMap := existingTopologyConfigMaps[0].(controller.Object).(*controller.RuntimeObject)
-	cmTopology := existingTopologyConfigMap.Object.(*corev1.ConfigMap)
-
-	if d, found := cmTopology.Data["topology"]; !found || strings.Compare(d, cm.Data["topology"]) != 0 {
-		_, err = r.Client.Resource(controller.ConfigMapsResource).Namespace(cm.Namespace).Update(ctx, unstructuredCM, metav1.UpdateOptions{})
-		if err != nil {
-			logger.Error(err, "failed to update topology configmap")
-		}
-		return err
-	}
-
-	return nil
-}
-
-type EventLogger struct{}
-
-func NewEventLogger() *EventLogger {
-	return &EventLogger{}
-}
-
-func (e *EventLogger) Log(ctx context.Context, resourceEvents []controller.ResourceEvent, _ *machinery.Topology, err error, _ *sync.Map) error {
-	logger := controller.LoggerFromContext(ctx).WithName("event logger")
-	for _, event := range resourceEvents {
-		// log the event
-		obj := event.OldObject
-		if obj == nil {
-			obj = event.NewObject
-		}
-		values := []any{
-			"type", event.EventType.String(),
-			"kind", obj.GetObjectKind().GroupVersionKind().Kind,
-			"namespace", obj.GetNamespace(),
-			"name", obj.GetName(),
-		}
-		if event.EventType == controller.UpdateEvent && logger.V(1).Enabled() {
-			values = append(values, "diff", cmp.Diff(event.OldObject, event.NewObject))
-		}
-		logger.Info("new event", values...)
-		if err != nil {
-			logger.Error(err, "error passed to reconcile")
-		}
-	}
-
-	return nil
 }
 
 // GetOldestKuadrant returns the oldest kuadrant resource from a list of kuadrant resources that is not marked for deletion.
